WIP: POST transaction

+++ src/main/java/br/com/guilherme/rinhabackend2024q1/dtos/TransactionDto.java
@@ -15,7 +15,7 @@ import org.hibernate.validator.constraints.Length;
 public class TransactionDto {
 
     @NotNull
-    private long valor;
+    private int valor;
     @NotNull
     private TransactionType tipo;
     @NotBlank
